move Program out of index
